# backend/static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DiabeCare API",
    "description": "DiabeCare \u2014 Diabetic Patient Medication Follow-Up System API",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "https://your-production-domain.com/api",
      "description": "Production"
    },
    {
      "url": "http://localhost:5000/api",
      "description": "Local dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/patients/": {
      "get": {
        "tags": [
          "Patients"
        ],
        "summary": "List patients \u2014 admins see all, doctors see their assigned patients, patients see only themselves",
        "responses": {
          "200": {
            "description": "Array of patient objects"
          }
        }
      },
      "post": {
        "tags": [
          "Patients"
        ],
        "summary": "Create patient record (admin/doctor)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "username",
                  "password",
                  "age",
                  "gender",
                  "diabetes_type",
                  "email"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "age": {
                    "type": "integer"
                  },
                  "gender": {
                    "type": "string",
                    "enum": [
                      "Male",
                      "Female"
                    ]
                  },
                  "diabetes_type": {
                    "type": "string",
                    "enum": [
                      "Type 1",
                      "Type 2",
                      "Gestational"
                    ]
                  },
                  "phone": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "example": "jdoe@email.com"
                  },
                  "blood_sugar": {
                    "type": "number"
                  },
                  "hba1c": {
                    "type": "number"
                  },
                  "assigned_doctor_id": {
                    "type": "integer"
                  },
                  "username": {
                    "type": "string",
                    "example": "jdoe"
                  },
                  "password": {
                    "type": "string",
                    "example": "pass1234"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Patient created"
          },
          "403": {
            "description": "Admin/doctor access required"
          }
        }
      }
    },
    "/patients/{patient_id}": {
      "get": {
        "tags": [
          "Patients"
        ],
        "summary": "Get single patient (includes medications)",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Patient object with medications array"
          },
          "404": {
            "description": "Patient not found"
          }
        }
      },
      "put": {
        "tags": [
          "Patients"
        ],
        "summary": "Update patient record",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "age": {
                    "type": "integer"
                  },
                  "gender": {
                    "type": "string",
                    "enum": [
                      "Male",
                      "Female"
                    ]
                  },
                  "diabetes_type": {
                    "type": "string",
                    "enum": [
                      "Type 1",
                      "Type 2",
                      "Gestational"
                    ]
                  },
                  "phone": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "blood_sugar": {
                    "type": "number"
                  },
                  "hba1c": {
                    "type": "number"
                  },
                  "last_visit": {
                    "type": "string",
                    "format": "date"
                  },
                  "next_visit": {
                    "type": "string",
                    "format": "date"
                  },
                  "username": {
                    "type": "string",
                    "description": "Updates the linked user account username"
                  },
                  "password": {
                    "type": "string",
                    "description": "Updates the linked user account password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Patient updated"
          },
          "404": {
            "description": "Patient not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Patients"
        ],
        "summary": "Delete patient (admin only)",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Patient deleted"
          },
          "403": {
            "description": "Admin access required"
          },
          "404": {
            "description": "Patient not found"
          }
        }
      }
    },
...
    "/patients/{patient_id}/reassign": {
      "put": {
        "summary": "Reassign a patient to a different doctor",
        "description": "Admin only. Sends SMS and in-app notifications to both patient and new doctor.",
        "tags": [
          "Patients"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "doctor_id"
                ],
                "properties": {
                  "doctor_id": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Patient reassigned successfully"
          },
          "403": {
            "description": "Admin only"
          },
          "404": {
            "description": "Patient or doctor not found"
          }
        }
      }
    },
    "/patients/reassign-bulk": {
      "put": {
        "summary": "Bulk reassign all patients from one doctor to another",
        "description": "Admin only. Moves all patients assigned to from_doctor_id to to_doctor_id. Sends SMS to each patient and a summary SMS to the new doctor.",
        "tags": [
          "Patients"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "from_doctor_id",
                  "to_doctor_id"
                ],
                "properties": {
                  "from_doctor_id": {
                    "type": "integer",
                    "description": "Doctor whose patients will be moved"
                  },
                  "to_doctor_id": {
                    "type": "integer",
                    "description": "Doctor who will receive the patients"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bulk reassignment complete with count"
          },
          "400": {
            "description": "Same doctor or missing fields"
          },
          "403": {
            "description": "Admin only"
          },
          "404": {
            "description": "Target doctor not found"
          }
        }
      }
    },
...
  },
...
}